Apply a S3 filter object to an ftmsData S3 object

This function takes a filter object of class 'moleculeFilt' and applies the filter to a dataset of class ftmsData

applyFilt(filter_object, ftmsObj, ...)

# S3 method for moleculeFilt
applyFilt(filter_object, ftmsObj, min_num = 2,
  ...)

# S3 method for massFilt
applyFilt(filter_object, ftmsObj, min_mass = 200,
  max_mass = 900, ...)

# S3 method for formulaFilt
applyFilt(filter_object, ftmsObj,
  remove = "NoFormula", ...)

# S3 method for emetaFilt
applyFilt(filter_object, ftmsObj, min_val = NULL,
  max_val = NULL, cats = NULL, na.rm = TRUE, ...)

Arguments

filter_object
an object of the class 'moleculeFilt' created by molecule_filter

ftmsObj
an object of the class ftmsData usually created by as.peakData

...
further arguments as described below based on the class of filter_object

min_num
[moleculeFilt object] an integer value specifying the minimum number of times each biomolecule must be observed across all samples in order to retain the biomolecule. Default value is 2.

min_mass
[massFilt object] a numeric value greater than 0, specifying the minimum mass a peak should have in order to retain the peak. Default value is 200.

max_mass
[massFilt object] a numeric value greater than min_mass, specifying the maximum mass a peak should have in order to retain the peak. Default value is 900.

remove
[formulaFilt object] a character string specifying which set of peaks to filter. Valid options are "NoFormula" and "Formula", defaults to "NoFormula".

min_val
[emetaFilt object] a numeric value specifying the minimum value (inclusive) that a peak should have for the specified 'e_meta' column.

max_val
[emetaFilt object] a numeric value specifying the maximum value (inclusive) that a peak should have for the specified 'e_meta' column.

cats
[emetaFilt object] a vector of character values specifying the level(s) of the specified 'e_meta' column which should be retained.

na.rm
[emetaFilt object] logical value specifying if peaks with NA values for the specified 'e_meta' column should be removed. Default value is TRUE.

Value

An object of the class ftmsData with specified molecules filtered out of the appropriate datasets.

See also