Create CoreMSData Object

Converts `CoreMSrbind` object to `CoreMSData` object

as.CoreMSData(
  all_data,
  index_cname = "Index",
  obs_mass_cname = "m/z",
  calib_mass_cname = "Calibrated m/z",
  calc_mass_cname = "Calculated m/z",
  pheight_cname = "Peak Height",
  error_cname = "Mass Error (ppm)",
  conf_cname = "Confidence Score",
  heteroatom_cname = "Heteroatom Class",
  iontype_cname = "Ion Type",
  file_cname = "Filename",
  monoiso_index_cname = "Mono Isotopic Index",
  mf_cname = "Molecular Formula",
  iso_cols = c("13C", "18O", "15N", "34S")
)

Arguments

all_data

a data.frame/CoreMSData object, should be the output of read_CoreMS_data

index_cname

a character string specifying the name of the column containing the index value

obs_mass_cname

a character string specifying the name of the column containing the observed mass to charge ratio (m/z) in all_data

calib_mass_cname

a character string specifying the name of the column containing the calibrated mass to charge ratio (m/z) in all_data

calc_mass_cname

a character string specifying the name of the column containing the calculated mass to charge ratio (m/z) in all_data

pheight_cname

a character string specifying the name of the column containing the peak height in all_data

error_cname

a character string specifying the name of the column containing the mass error (ppm) in all_data

conf_cname

a character string specifying the name of the column containing the confidence score in all_data

heteroatom_cname

a character string specifying the name of the column containing the heteroatom class in all_data

iontype_cname

a characer string specifying the name of the column containing the ion type in all_data

file_cname

a character string specifying the name of the column containing the filenames in all_data

monoiso_index_cname

a character string specifying the name of the column containing the monoisotopic index in all_data

mf_cname

a character string specifying the name of the column containing the molecular formula in all_data

iso_cols

a character vector specifying columns indicating the presence of isotopic peaks, e.g. a column "13C" with T/F specifying the presence of 13C. Defaults to c("13C", "18O", "15N", "34S").

Value

`CoreMSData` object, consisting of a list of two data frames

Author

Natalie Winans