Convert Data to reactionData Class

Converts a list object or several data.frames of FT-MS data to an object of the class 'reactionData'. Objects of the class 'reactionData' are lists with two obligatory components e_data and f_data and one optional compounent, e_meta.

as.reactionData(
  e_data,
  f_data,
  e_meta = NULL,
  edata_cname,
  fdata_cname,
  reaction_cname,
  instrument_type = "12T",
  db = NA,
  ...
)

Arguments

e_data

a \(p \times n + 1\) data.frame of expression data, where \(p\) is the number of observed reactions and \(n\) is the number of samples. Each row corresponds to data for each peak. One column specifying a unique identifier for each peak/mass (row) must be present.

f_data

a data.frame with \(n\) rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.

e_meta

a data.frame with \(p\) rows. Each row corresponds to a reaction with one column giving a unique reaction identifier (must be named the same as the column in e_data) and other columns giving meta information. At a minimum a column giving the mass of each peak and a column giving molecular formulae or columns giving elemental counts must be present.

edata_cname

character string specifying the name of the column containing a unique identifier for each reaction in e_data and e_meta.

fdata_cname

character string specifying the name of the column containing the sample identifiers in f_data.

reaction_cname

character string specifying the name of the column containing the reaction identifiers in e_meta.

instrument_type

character string giving the type of FT-MS instrument data was generated by. Valid options are: "12T" and "21T". Defaults to "12T". This information is used to determine appropriate plotting functions for Van Krevelen, Kendrick, etc. plots.

db

character string specifying the database from which reaction information is drawn

...

further arguments

Details

as.reactionData constructs a reactionData object which is an ftmsData object where the rows of e_data correspond to reactions.