FTICR Peak Meta Information

A dataset containing the meta information associated with each peak.

Format

A data.frame with 2,442 rows (peaks/masses) and 10 columns:

Mass

Mass identifier (matches Mass column in edata) - negative ion mass as samples were analyzed by electrospray ionization in negative mode

C

Number of Carbons making up identified compound

H

Number of Hydrogens making up identified compound

O

Number of Oxygens making up identified compound

N

Number of Nitrogens making up identified compound

C13

Indicates if compound contains Carbon13 if value is '1'

S

Number of Sulfurs making up identified compound

P

Number of Phosphorus making up identified compound

Error

Difference between the experimental mass and the accurate mass

NeutralMass

Experimental mass + mass of hydrogen - mass of electron