Kendrick plot of an ftmsData object

Constructs a Kendrick plot (Kendrick defect vs Kendrick mass) for an object of type ftmsData.

kendrickPlot(
  ftmsObj,
  title = NA,
  colorPal = NA,
  colorCName = NA,
  vkBoundarySet = "bs1",
  xlabel = "Kendrick Mass",
  ylabel = "Kendrick Defect",
  legendTitle = colorCName,
  base_compound = NULL
)

Arguments

ftmsObj

an object of class 'peakData' or 'compoundData', typically a result of as.peakData or mapPeaksToCompounds.

title

plot title

colorPal

color palette function, one of col_numeric, col_factor or similar scales palette function

colorCName

column name of ftmsObj$e_meta to use for coloring the points. If NA and vkBoundarySet is provided, the points will be colored according to the Van Krevelen category.

vkBoundarySet

character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of assign_class for differences in sets.

xlabel

x axis label, default is "Kendrick Mass"

ylabel

y axis label, default is "Kendrick Defect"

legendTitle

title for the legend, only used when coloring points according to a numeric scale

base_compound

string specifying the base compound used to calculate the Kendrick Mass you want plotted on the x-axis. Defaults to NULL, in which case the first Kendrick Mass is used.

Value

a plotly object

See also

Author

Amanda White