Generates mass-error plot for CoreMS output data

Generates mass-error plot for CoreMS output data

mass_error_plot(
  cmsObj,
  min_conf = NULL,
  title = NULL,
  xlabel = "Mass to charge ratio (m/z)",
  ylabel = "Mass error (ppm)",
  xrange = NULL,
  yrange = NULL,
  color_pal = NULL,
  log_color_scale = FALSE
)

Arguments

cmsObj

a data.frame of MS data with columns containing calculated m/z, mass error (ppm), and file names, the output of function read_CoreMS_data

min_conf

a numeric value between 0 and 1 specifying a minimum confidence threshold to apply to the dataset. Peaks with confidence scores lower than the threshold will be omitted from the plot.

title

a character string specifying the plot title

xlabel

a character string specifying the x-axis label

ylabel

a character string specifying the y-axis label

xrange

a numerical vector of length 2 specifying the x-axis range

yrange

a numerical vector of length 2 specifying the y-axis range

color_pal

has not yet been implemented

log_color_scale

logical, if TRUE, color scale will be log-transformed. Defaults to FALSE.

Value

plotly object