Summary method for moleculeFilt object

Summary method for moleculeFilt object

# S3 method for moleculeFilt
summary(object, min_num = NULL, ...)

Arguments

object

an object of class 'moleculeFilt' created by running molecule_filter

min_num

an integer value specifying the minimum number of times each peak must be observed across all samples

...

included for compliance with generic method

Value

If min_num is provided, a summary of the effect of implementing a filter with the specified threshold. Otherwise, a summary of the number of peaks which were observed over the number of possible samples.