Calculate and Count Observed Mass Transformations

Calculate and Count Observed Mass Transformations

transformation_counts(ftmsObj, transformDF, transformDigits = 4,
  transMass_cname, transID_cname, transOther_cname = NULL)

Arguments

ftmsObj
an object of class 'peakData' or 'compoundData', typically a result of as.peakData or mapPeaksToCompounds.

transformDF
a data.frame of known transformations of interest, with a minimum of two columns (in any order): column 'transMass_cname' gives the transformation masses which should be counted and column 'transID_cname' gives a unique identifier (usually character string) for each transformation mass

transformDigits
the number of decimal places that should be retained for the purposes of matching observed transformations to transformations in transformDF. Defaults to 4

transMass_cname
a character string giving the name of the column in transformDF which gives the mass of the transformations of interest

transID_cname
a character string giving the name of the column in transformDF which gives a unique identifier for each transformation of interest

transOther_cname
a character vector specifying any other columns in transformDF which should be returned with the transformation counts. This is only applicable if transformDF has more than 2 columns with information that should be carried through to results. Defaults to NULL

Value

a data.frame with counts of the the number of times that each transformation in transformDF was observed in each sample