Van Krevelen plot of an ftmsData object

Create a Van Krevelen plot (H:C ratio vs O:C ratio) for an ftmsData object using Plotly.

vanKrevelenPlot(
  ftmsObj,
  title = NA,
  colorPal = NA,
  colorCName = NA,
  vkBoundarySet = "bs1",
  showVKBounds = TRUE,
  xlabel = "O:C Ratio",
  ylabel = "H:C Ratio",
  legendTitle = colorCName
)

Arguments

ftmsObj

an object of class 'peakData' or 'compoundData', (output of as.peakData or mapPeaksToCompounds), or a 'groupSummary' object (output of summarizeGroups) or a 'comparisonSummary' object (output of summarizeGroupComparisons).

title

plot title

colorPal

color palette function, one of col_numeric, col_factor or similar scales palette function

colorCName

column name of ftmsObj$e_data or ftmsObj$e_meta to use for coloring the points. If NA and vkBoundarySet is provided, the points will be colored according to the Van Krevelen category.

vkBoundarySet

character vector specifying which boundary set to use when determining class. Valid options are currently "bs1" and "bs2" and defaults to "bs1". See details of assign_class for differences in sets.

showVKBounds

TRUE/FALSE, should the Van Krevelen bounds be plotted?

xlabel

x axis label, default is "O:C Ratio"

ylabel

y axis label, default is "H:C Ratio"

legendTitle

title for the legend, only used when coloring points according to a numeric scale

Value

a plotly object

See also

Author

Amanda White