Convenience function to wrap ftmsRanalysis plotting functions for Trelliscope

Convenience function to wrap ftmsRanalysis plotting functions for Trelliscope

panelFunctionGenerator(plot_fn_name, ...)

Arguments

plot_fn_name
plot function name, e.g. "vanKrevelenPlot" or "kendrickPlot"

...
other parameters to pass to the plotting function other than the data object

Details

Sometimes the additional parameters to pass to the plotting function will depend on the key associated with the data in the ddo (e.g. if a column name is prepended with the group name). In that case, see example below for how to construct an expr statement to obtain the necessary information.

Examples

library(ftmsRanalysis)
library(trelliscope)

vdbDir <- vdbConn(file.path(tempdir(), "trell_test"), autoYes = TRUE) #temporary directory

#> *** Copying latest viewer to vdb directory...

data('exampleProcessedPeakData')
groupDdo <- divideByGroup(exampleProcessedPeakData)
groupSummaryDdo <- summarizeGroups(groupDdo, summary_functions = c("prop_present", "n_present"))

#> *** finding global variables used in 'fn'...

#> 
#>   found: summary_functions

#> *** testing 'fn' on a subset...

#>  ok

#> * Applying recombination...

# See expr here:
panelFnG1 <- panelFunctionGenerator("vanKrevelenPlot", colorCName=expr(paste0(getSplitVar(v, "Group"), "_prop_present")),
                                    legendTitle="Proportion<br>Present")

#> Error in expr(paste0(getSplitVar(v, "Group"), "_prop_present")): could not find function "expr"

makeDisplay(groupSummaryDdo,
            panelFn=panelFnG1,
            name = "Trelliscope test G_1 with VK plot per group",
            group = "Group")

#> * Validating 'panelFn'...

#> Error in formals(fn): object 'panelFnG1' not found

view()

#> Error: Visualization database does not have any displays: /private/var/folders/5p/fg3v33qx4ks7wr2rhchln2jhpsswgc/T/RtmpZmw3rr/trell_test